some fix

+++ src/main/java/com/tabletka/service/OrderService.java
@@ -2,7 +2,6 @@ package com.tabletka.service;
 
 import com.tabletka.exception.UserIsNotLoggedInException;
 import com.tabletka.model.order.Order;
-import com.tabletka.model.pharmacy.Pharmacy;
 
 import java.util.List;
 
@@ -11,5 +10,5 @@ public interface OrderService {
 
     List<Order> getOrdersForPharmacy(Long pharmacyId);
 
-    void makeOrder(Long ProductId, int amount) throws UserIsNotLoggedInException;
+    void makeOrder(Long ProductId, Long amount) throws UserIsNotLoggedInException;
 }
